+++ soh/soh/Enhancements/Difficulty/PermanentLosses.cpp
@@ -1,3 +1,6 @@
+#include <ship/Context.h>
+#include <ship/window/Window.h>
+
 #include "soh/Enhancements/game-interactor/GameInteractor_Hooks.h"
 #include "soh/Enhancements/randomizer/randomizer.h"
 #include "soh/OTRGlobals.h"
