+++ soh/soh/Enhancements/randomizer/SeedContext.h
@@ -2,7 +2,6 @@
 
 #include "randomizerTypes.h"
 #include "z64save.h"
-#include "item_location.h"
 #include "item_override.h"
 #include "soh/Enhancements/custom-message/text.h"
 #include "hint.h"
@@ -33,6 +32,7 @@ class DungeonInfo;
 class TrialInfo;
 class Trials;
 class Kaleido;
+class ItemLocation;
 
 class Context {
   public:
