+++ soh/soh/ShipUtils.h
@@ -1,10 +1,11 @@
 #ifndef SHIP_UTILS_H
 #define SHIP_UTILS_H
 
-#include <libultraship/libultraship.h>
+#include <libultraship/libultra.h>
 
 #ifdef __cplusplus
 #include <set>
+#include <vector>
 
 void LoadGuiTextures();
