+++ soh/soh/SohGui/Menu.h
@@ -2,7 +2,7 @@
 #define MENU_H
 
 #include <libultraship/libultraship.h>
-#include <fast/backends/gfx_rendering_api.h>
+#include <fast/Fast3dWindow.h>
 #include "MenuTypes.h"
 
 namespace Ship {
